coded-be/requirements.txt
django>=6.0.3
djangorestframework>=3.17.1
djangorestframework-simplejwt
Pillow
django-cors-headers
